更新操作修改为全量更新

+++ learning-initial/src/main/resources/init.sql
@@ -4,7 +4,7 @@
 -- ----------------------------
 DROP TABLE IF EXISTS `sys_acl`;
 CREATE TABLE `sys_acl` (
-  `id` int(11) NOT NULL AUTO_INCREMENT COMMENT '权限id',
+  `id` bigint(11) NOT NULL AUTO_INCREMENT COMMENT '权限id',
   `code` varchar(20) NOT NULL DEFAULT '' COMMENT '权限码',
   `name` varchar(20) NOT NULL DEFAULT '' COMMENT '权限名称',
   `acl_module_id` int(11) NOT NULL DEFAULT '0' COMMENT '权限所在的权限模块id',
@@ -25,9 +25,9 @@ CREATE TABLE `sys_acl` (
 -- ----------------------------
 DROP TABLE IF EXISTS `sys_acl_module`;
 CREATE TABLE `sys_acl_module` (
-  `id` int(11) NOT NULL AUTO_INCREMENT COMMENT '权限模块id',
+  `id` bigint(11) NOT NULL AUTO_INCREMENT COMMENT '权限模块id',
   `name` varchar(20) NOT NULL DEFAULT '' COMMENT '权限模块名称',
-  `parent_id` int(11) NOT NULL DEFAULT '0' COMMENT '上级权限模块id',
+  `parent_id` bigint(11) NOT NULL DEFAULT '0' COMMENT '上级权限模块id',
   `level` varchar(200) NOT NULL DEFAULT '' COMMENT '权限模块层级',
   `seq` int(11) NOT NULL DEFAULT '0' COMMENT '权限模块在当前层级下的顺序，由小到大',
   `status` int(11) NOT NULL DEFAULT '1' COMMENT '状态，1：正常，0：冻结',
@@ -44,9 +44,9 @@ CREATE TABLE `sys_acl_module` (
 -- ----------------------------
 DROP TABLE IF EXISTS `sys_dept`;
 CREATE TABLE `sys_dept` (
-  `id` int(11) NOT NULL AUTO_INCREMENT COMMENT '部门id',
+  `id` bigint(11) NOT NULL AUTO_INCREMENT COMMENT '部门id',
   `name` varchar(20) NOT NULL DEFAULT '' COMMENT '部门名称',
-  `parent_id` int(11) NOT NULL DEFAULT '0' COMMENT '上级部门id',
+  `parent_id` bigint(11) NOT NULL DEFAULT '0' COMMENT '上级部门id',
   `level` varchar(200) NOT NULL DEFAULT '' COMMENT '部门层级',
   `seq` int(11) NOT NULL DEFAULT '0' COMMENT '部门在当前层级下的顺序，由小到大',
   `remark` varchar(200) DEFAULT '' COMMENT '备注',
@@ -62,9 +62,9 @@ CREATE TABLE `sys_dept` (
 -- ----------------------------
 DROP TABLE IF EXISTS `sys_log`;
 CREATE TABLE `sys_log` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(11) NOT NULL AUTO_INCREMENT,
   `type` int(11) NOT NULL DEFAULT '0' COMMENT '权限更新的类型，1：部门，2：用户，3：权限模块，4：权限，5：角色，6：角色用户关系，7：角色权限关系',
-  `target_id` int(11) NOT NULL COMMENT '基于type后指定的对象id，比如用户、权限、角色等表的主键',
+  `target_id` bigint(11) NOT NULL COMMENT '基于type后指定的对象id，比如用户、权限、角色等表的主键',
   `old_value` text COMMENT '旧值',
   `new_value` text COMMENT '新值',
   `operator` varchar(20) NOT NULL DEFAULT '' COMMENT '操作者',
@@ -80,7 +80,7 @@ CREATE TABLE `sys_log` (
 -- ----------------------------
 DROP TABLE IF EXISTS `sys_role`;
 CREATE TABLE `sys_role` (
-  `id` int(11) NOT NULL AUTO_INCREMENT COMMENT '角色id',
+  `id` bigint(11) NOT NULL AUTO_INCREMENT COMMENT '角色id',
   `name` varchar(20) NOT NULL,
   `type` int(11) NOT NULL DEFAULT '1' COMMENT '角色的类型，1：管理员角色，2：其他',
   `status` int(11) NOT NULL DEFAULT '1' COMMENT '状态，1：可用，0：冻结',
@@ -97,9 +97,9 @@ CREATE TABLE `sys_role` (
 -- ----------------------------
 DROP TABLE IF EXISTS `sys_role_acl`;
 CREATE TABLE `sys_role_acl` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `role_id` int(11) NOT NULL COMMENT '角色id',
-  `acl_id` int(11) NOT NULL COMMENT '权限id',
+  `id` bigint(11) NOT NULL AUTO_INCREMENT,
+  `role_id` bigint(11) NOT NULL COMMENT '角色id',
+  `acl_id` bigint(11) NOT NULL COMMENT '权限id',
   `operator` varchar(20) NOT NULL DEFAULT '' COMMENT '操作者',
   `operate_time` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '最后一次更新的时间',
   `operate_ip` varchar(200) NOT NULL DEFAULT '' COMMENT '最后一次更新者的ip',
@@ -112,9 +112,9 @@ CREATE TABLE `sys_role_acl` (
 -- ----------------------------
 DROP TABLE IF EXISTS `sys_role_user`;
 CREATE TABLE `sys_role_user` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `role_id` int(11) NOT NULL COMMENT '角色id',
-  `user_id` int(11) NOT NULL COMMENT '用户id',
+  `id` bigint(11) NOT NULL AUTO_INCREMENT,
+  `role_id` bigint(11) NOT NULL COMMENT '角色id',
+  `user_id` bigint(11) NOT NULL COMMENT '用户id',
   `operator` varchar(20) NOT NULL DEFAULT '' COMMENT '操作者',
   `operate_time` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '最后一次更新的时间',
   `operate_ip` varchar(20) NOT NULL DEFAULT '' COMMENT '最后一次更新者的ip地址',
@@ -127,12 +127,12 @@ CREATE TABLE `sys_role_user` (
 -- ----------------------------
 DROP TABLE IF EXISTS `sys_user`;
 CREATE TABLE `sys_user` (
-  `id` int(11) NOT NULL AUTO_INCREMENT COMMENT '用户id',
+  `id` bigint(11) NOT NULL AUTO_INCREMENT COMMENT '用户id',
   `username` varchar(20) NOT NULL DEFAULT '' COMMENT '用户名称',
   `telephone` varchar(13) NOT NULL DEFAULT '' COMMENT '手机号',
   `mail` varchar(20) NOT NULL DEFAULT '' COMMENT '邮箱',
   `password` varchar(40) NOT NULL DEFAULT '' COMMENT '加密后的密码',
-  `dept_id` int(11) NOT NULL DEFAULT '0' COMMENT '用户所在部门的id',
+  `dept_id` bigint(11) NOT NULL DEFAULT '0' COMMENT '用户所在部门的id',
   `status` int(11) NOT NULL DEFAULT '1' COMMENT '状态，1：正常，0：冻结状态，2：删除',
   `remark` varchar(200) DEFAULT '' COMMENT '备注',
   `operator` varchar(20) NOT NULL DEFAULT '' COMMENT '操作者',
